+++ jdaviz/configs/default/plugins/data_quality/data_quality.py
@@ -1,3 +1,4 @@
+from itertools import chain
 import os
 from traitlets import Any, Dict, Bool, List, Unicode, Float, observe
 
@@ -10,11 +11,12 @@ from jdaviz.core.registries import tray_registry
 from jdaviz.core.template_mixin import (
     PluginTemplateMixin, LayerSelect, ViewerSelectMixin
 )
-from jdaviz.core.user_api import PluginUserApi
 from jdaviz.core.tools import ICON_DIR
+from jdaviz.core.user_api import PluginUserApi
 from jdaviz.configs.default.plugins.data_quality.dq_utils import (
     decode_flags, generate_listed_colormap, dq_flag_map_paths, load_flag_map
 )
+
 
 __all__ = ['DataQuality']
 
@@ -28,7 +30,7 @@ telescope_names = {
 class DataQuality(PluginTemplateMixin, ViewerSelectMixin):
     template_file = __file__, "data_quality.vue"
 
-    irrelevant_msg = Unicode("Data Quality plugin is in development.").tag(sync=True)
+    irrelevant_msg = Unicode("").tag(sync=True)
 
     # `layer` is the science data layer
     science_layer_multiselect = Bool(False).tag(sync=True)
@@ -55,12 +57,12 @@ class DataQuality(PluginTemplateMixin, ViewerSelectMixin):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-
         self.icons = {k: v for k, v in self.app.state.icons.items()}
 
         self.science_layer = LayerSelect(
             self, 'science_layer_items', 'science_layer_selected',
-            'viewer_selected', 'science_layer_multiselect', is_root=True
+            'viewer_selected', 'science_layer_multiselect',
+            is_root=True, has_children=True
         )
 
         self.dq_layer = LayerSelect(
@@ -71,6 +73,40 @@ class DataQuality(PluginTemplateMixin, ViewerSelectMixin):
 
         self.load_default_flag_maps()
         self.init_decoding()
+        self._set_irrelevant()
+        self._update_available_viewers()
+
+    def _update_available_viewers(self):
+        if not hasattr(self, 'viewer'):
+            return
+
+        def is_image_viewer(viewer):
+            from jdaviz.configs.imviz.plugins.viewers import ImvizImageView
+            from jdaviz.configs.cubeviz.plugins.viewers import CubevizImageView
+
+            return isinstance(viewer, (ImvizImageView, CubevizImageView))
+
+        viewer_filter_names = [filt.__name__ for filt in self.viewer.filters]
+        if 'is_image_viewer' not in viewer_filter_names:
+            self.viewer.add_filter(is_image_viewer)
+
+        # by default, select all image viewers to sync each DQ layer's
+        # options across all viewers that they're visible in:
+        self.viewer_multiselect = True
+        self.viewer.select_all()
+
+    @observe('dq_layer_items')
+    def _set_irrelevant(self, *args):
+        self._update_available_viewers()
+
+        children_available = any([
+            len(assoc['children']) > 0
+            for label, assoc in getattr(self.app, '_data_associations', {}).items()
+        ])
+        self.irrelevant_msg = (
+            '' if children_available else
+            "No Data Quality layers available."
+        )
 
     @observe('science_layer_selected')
     def update_dq_layer(self, *args):
@@ -91,10 +127,22 @@ class DataQuality(PluginTemplateMixin, ViewerSelectMixin):
             self.flag_map_items = self.flag_map_items + [telescope_names[name]]
 
     @property
-    def unique_flags(self):
+    def dq_layer_selected_flattened(self):
         selected_dq = self.dq_layer.selected_obj
         if not len(selected_dq):
             return []
+
+        if isinstance(selected_dq, list) and isinstance(selected_dq[0], list):
+            # flatten a nested list:
+            selected_dq = list(chain.from_iterable(selected_dq))
+        elif not isinstance(selected_dq, list):
+            # if it's a single layer, make it a list:
+            selected_dq = [selected_dq]
+        return selected_dq
+
+    @property
+    def unique_flags(self):
+        selected_dq = self.dq_layer_selected_flattened
 
         dq = selected_dq[0].get_image_data()
         return np.unique(dq[~np.isnan(dq)])
@@ -112,7 +160,7 @@ class DataQuality(PluginTemplateMixin, ViewerSelectMixin):
         self.flag_map_definitions_selected = selected
 
     @observe('dq_layer_selected')
-    def init_decoding(self, event={}):
+    def init_decoding(self, event={}, viewers=None):
         if not self.validate_flag_decode_possible:
             return
 
@@ -123,54 +171,84 @@ class DataQuality(PluginTemplateMixin, ViewerSelectMixin):
             unique_flags=unique_flags,
             rgba_colors=rgba_colors
         )
-        dq_layer = self.get_dq_layer()
-        dq_layer.composite._allow_bad_alpha = True
+        dq_layers = self.get_dq_layers(viewers=viewers)
 
-        flag_bits = np.array([flag['flag'] for flag in self.decoded_flags])
+        for dq_layer in dq_layers:
+            dq_layer.composite._allow_bad_alpha = True
 
-        dq_layer.state.stretch = 'lookup'
-        stretch_object = dq_layer.state.stretch_object
-        stretch_object.flags = flag_bits
+            # for cubeviz, also change uncert-viewer defaults to
+            # map the out-of-bounds regions to the cmap's `bad` color:
+            if self.app.config == 'cubeviz':
+                uncert_viewer = self.app.get_viewer(
+                    self.app._jdaviz_helper._default_uncert_viewer_reference_name
+                )
+                for layer in uncert_viewer.layers:
+                    layer.composite._allow_bad_alpha = True
+                    layer.force_update()
 
-        with delay_callback(dq_layer.state, 'alpha', 'cmap', 'v_min', 'v_max'):
-            if len(flag_bits):
-                dq_layer.state.v_min = min(flag_bits)
-                dq_layer.state.v_max = max(flag_bits)
+            flag_bits = np.array([flag['flag'] for flag in self.decoded_flags])
 
-            dq_layer.state.alpha = self.dq_layer_opacity
-            dq_layer.state.cmap = cmap
+            dq_layer.state.stretch = 'lookup'
+            stretch_object = dq_layer.state.stretch_object
+            stretch_object.flags = flag_bits
 
-    def get_dq_layer(self):
+            with delay_callback(dq_layer.state, 'alpha', 'cmap', 'v_min', 'v_max'):
+                if len(flag_bits):
+                    dq_layer.state.v_min = min(flag_bits)
+                    dq_layer.state.v_max = max(flag_bits)
+
+                dq_layer.state.alpha = self.dq_layer_opacity
+                dq_layer.state.cmap = cmap
+
+    def get_dq_layers(self, viewers=None):
         if self.dq_layer_selected == '':
             return
 
-        viewer = self.viewer.selected_obj
-        [dq_layer] = [
-            layer for layer in viewer.layers if
+        if viewers is None:
+            viewers = self.viewer.selected_obj
+
+        if not hasattr(viewers, '__len__'):
+            viewers = [viewers]
+
+        dq_layers = [
+            layer for viewer in viewers
+            for layer in viewer.layers if
             layer.layer.label == self.dq_layer_selected
         ]
-        return dq_layer
+        return dq_layers
 
-    def get_science_layer(self):
-        viewer = self.viewer.selected_obj
-        [science_layer] = [
-            layer for layer in viewer.layers if
+    def get_science_layers(self, viewers=None):
+
+        if viewers is None:
+            viewers = self.viewer.selected_obj
+
+        if not hasattr(viewers, '__len__'):
+            viewers = [viewers]
+
+        science_layers = [
+            layer for viewer in viewers
+            for layer in viewer.layers if
             layer.layer.label == self.science_layer_selected
         ]
-        return science_layer
+        return science_layers
 
     @observe('dq_layer_opacity')
     def update_opacity(self, event={}):
-        science_layer = self.get_science_layer()
-        dq_layer = self.get_dq_layer()
+        science_layers = self.get_science_layers()
+        selected_dq = self.dq_layer_selected_flattened
 
-        if dq_layer is not None:
-            # DQ opacity is a fraction of the science layer's opacity:
-            dq_layer.state.alpha = self.dq_layer_opacity * science_layer.state.alpha
+        if len(selected_dq):
+            for sci_layer, dq_layer in zip(science_layers, selected_dq):
+                # DQ opacity is a fraction of the science layer's opacity:
+                dq_layer.state.alpha = self.dq_layer_opacity * sci_layer.state.alpha
 
     @observe('decoded_flags', 'flags_filter')
-    def _update_cmap(self, event={}):
-        dq_layer = self.get_dq_layer()
+    def _update_cmap(self, event={}, viewers=None):
+        dq_layers = self.get_dq_layers(viewers=viewers)
+
+        if dq_layers is None:
+            return
+
         flag_bits = np.array([flag['flag'] for flag in self.decoded_flags])
         rgb_colors = [hex2color(flag['color']) for flag in self.decoded_flags]
 
@@ -190,28 +268,31 @@ class DataQuality(PluginTemplateMixin, ViewerSelectMixin):
             )
         ])
 
-        with delay_callback(dq_layer.state, 'v_min', 'v_max', 'alpha', 'stretch', 'cmap'):
-            # set correct stretch and limits:
-            # dq_layer.state.stretch = 'lookup'
-            stretch_object = dq_layer.state.stretch_object
-            stretch_object.flags = flag_bits
-            stretch_object.dq_array = dq_layer.get_image_data()
-            stretch_object.hidden_flags = hidden_flags
+        for dq_layer in dq_layers:
+            with delay_callback(
+                    dq_layer.state, 'v_min', 'v_max', 'alpha', 'stretch', 'cmap'
+            ):
+                # set correct stretch and limits:
+                # dq_layer.state.stretch = 'lookup'
+                stretch_object = dq_layer.state.stretch_object
+                stretch_object.flags = flag_bits
+                stretch_object.dq_array = dq_layer.get_image_data()
+                stretch_object.hidden_flags = hidden_flags
 
-            # update the colors of the listed colormap without
-            # reassigning the layer.state.cmap object
-            cmap = dq_layer.state.cmap
-            cmap.colors = rgb_colors
-            cmap._init()
+                # update the colors of the listed colormap without
+                # reassigning the layer.state.cmap object
+                cmap = dq_layer.state.cmap
+                cmap.colors = rgb_colors
+                cmap._init()
 
-            # trigger updates to cmap in viewer:
-            dq_layer.update()
+                # trigger updates to cmap in viewer:
+                dq_layer.update()
 
-            if len(flag_bits):
-                dq_layer.state.v_min = min(flag_bits)
-                dq_layer.state.v_max = max(flag_bits)
+                if len(flag_bits):
+                    dq_layer.state.v_min = min(flag_bits)
+                    dq_layer.state.v_max = max(flag_bits)
 
-            dq_layer.state.alpha = self.dq_layer_opacity
+                dq_layer.state.alpha = self.dq_layer_opacity
 
     def update_visibility(self, index):
         self.decoded_flags[index]['show'] = not self.decoded_flags[index]['show']
@@ -237,15 +318,20 @@ class DataQuality(PluginTemplateMixin, ViewerSelectMixin):
         if not hasattr(self, 'science_layer'):
             return
 
-        layer = self.science_layer.selected_obj
+        layer = self.get_science_layers()
         if not len(layer):
             return
 
         # this is defined for JWST and ROMAN, should be upper case:
         telescope = layer[0].layer.meta.get('telescope', None)
 
+        if telescope is None:
+            # for spectral cubes in Cubeviz:
+            telescope = layer[0].layer.meta.get('_primary_header', {}).get('TELESCOP', None)
+
         if telescope is not None:
-            self.flag_map_selected = telescope_names[telescope.lower()]
+            flag_map_to_select = telescope_names.get(telescope.lower())
+            self.flag_map_selected = flag_map_to_select
 
     def vue_hide_all_flags(self, event):
         for flag in self.decoded_flags:
@@ -271,7 +357,7 @@ class DataQuality(PluginTemplateMixin, ViewerSelectMixin):
             expose=(
                 'science_layer', 'dq_layer',
                 'decoded_flags', 'flags_filter',
-                'viewer', 'dq_layer_opacity',
+                'dq_layer_opacity',
                 'flag_map_definitions_selected',
             )
         )